feat: put presentation window in full screen on second monitor
If there is only one monitor, then the presentation window will go
fullscreen on that only monitor. In the cas we have multiple monitors,
we simply put the presentation window on the second monitor.
This will have to be reworked in order to propose to which monitor to
display (and go fullscreen) for the presentation window.